copy from mkiocccentry dyn_array subdirectory
This repo is a copy mkiocccentry `dyn_array` version "2.3 2023-02-04"
and `dyn_test` version "1.9 2023-02-04".

The `README.md` file was modified to have better markdown header levels.
Moreover a Source Code History section was added to detail where
this code came from as well as a detailed history of GitHub repo
commits as well as dates.  This file also states the pre-git
origin of this code.

The `.gitignore` file was added.