Only define parse_verbosity() in some conditions
This change is necessary because in this repo we do not have access to
parse_verbosity() but in the dbg() repo we do (it's defined in dbg.[ch])
and in the mkiocccentry repo we also do. Thus if DBG_USE is not defined
we include the function. In dyn_test.c the macro is not defined so it is
provided in the file.