Rename dyn_array.a libdyn_array.a
To allow for linking in this library from say /usr/local/lib the library
file has to be renamed to /usr/local/libdyn_array.a. This is a linker
implementation detail. We need to have the ability to link in the dbg
API in the jparse repo (https://github.com/xexyl/jparse) and this now is
easier to link in and thus use the library.

The README.md set up instructions have two ways now: first what already
existed, the set up of the library if you do not wish to install it and
now the second, how to do it if you wish to install it.

I note that the Makefile here still refers to the dbg repo as ../dbg but
it is not clear to me if this should be changed to a system install. If
so that might necessitate updating the dbg README.md file (or maybe not
.. not sure).