Fix Makefile for libdbg.a
As the dbg repo is required the Makefile now links in libdbg.a rather
than try and do ../dbg/dbg.a (an oversight that was missed in
c8144fb58bc5675ca1dd558d3537473cfb0deb3f).

The README.md now states that one needs the dbg API installed in order
to use this repo and how one might obtain it, compile and install it.