Add to Makefile -L../dbg for mkiocccentry
This is unfortunate but in order for the self-contained dyn_array within
the mkiocccentry repo (thus not requiring the dyn_array repo or dbg repo
to be installed) we have to have -L../dbg for the dbg library in the
dyn_array Makefile in the mkiocccentry repo. This means that this repo
has to have it too.