eliminate the need for dbg to be installed to compile and test
The dbg library and header files are no longer required to be
installed in order to compile this code and to test this code.

The `dyn_test.c` file that was responsible for testing
the dyn_array code, contains internal private symbols,
variables and functions that eliminate the need to
have the dbg library and header files installed just
to compile and test the dyn_array code.

Updated `README.md` to account for the above.
Added suggestions to test dyn_array after compiling.
Changed note to just say that dbg must be installed
to use the dyn_array library.

Removed any reference to dbg from `Makefile` other
than to force DBG_USE to be undefined during the compiling
of the `dyn_test.c` file.

Changed `DYN_TEST_VERSION` from "1.10.1 2024-08-28"
to "1.11 2024-08-31".

IMPORTANT NOTE: There was **NO** change to the dyn_array API.
The `dyn_array.c` and `dyn_array.h` were **NOT** touched.
The contents of `libdyn_array.a` was **NOT** changed.
These changes are only for the convenience of compiling dyn_array.

Performed `make clobber depend all test` to test the above.