improve make depend
We remove all `/usr/include/` and `/usr/local/include/` files
form dependency rules by applying the `sed(1)` rules with "g".

This did not change the dependencies for this `Makefile`,
nevertheless we apply the above fix anyway in case it is
ever needed "down the road" as the expression goes.

Performed `make clobber depend all test` to test the above.