Improve Makefile (install, uninstall)
Add PREFIX variable to Makefile to allow for installing to a different
location than the default /usr/local.

Add uninstall rule.

To be congruent with the install rule the uninstall rule has a way to
silence the -v option to rm as well (instead of the install command).

Updated README.md. I note that the structure of the file could be
improved upon but I do not have time for this right now.