fix .github/workflows/test.yml
We removed processing with shellcheck since there are
no shell scripts, nor does `make shellcheck` do anything useful.

We removed processing with seqcexit since this is a
development process, not a testing method.  Failure to
perform seqcexit is not a testing function.

We clone and install dbg.

We added `make picky` to the set.

We added `make check_man` to the set.

We perform `make all` before doing just `make`.

Performed `make clobber depend all test` to test the above.