Fix rm in Makefile rules
Add where missing ${I} to ${RM} lines.

Change where it was not ${I} at the beginning of ${RM} lines.

Add the ${RM_V} to the ${RM} lines where it was missing (it is
analogous to ${INSTALL_V} but the default of -v could certainly be
changed if it is a problem).