Fix make depend
As a series of commands the ${RM} cannot have the ${Q} in it; that is
for the start of the set of commands.

Also since the ${RM} did not in the past use -v set the ${RM_V} to empty
string by default.

What would have detected this earlier is for a make prep/release rule
like some other repos have but at this point one does not exist. Instead
I have put this into the test workflow file.