make cleaning, clobbering, uninstalling more verbose
When doing a `make clean` (and/or `make legacy_clean`), a `make clobber`
(and/or a  `make legacy_lobber`), or a `make uninstall` is it better to
inform the user what you are doing than to silently remove stuff.

And since one is needing to echo removal related commands, then one might
we well be more informative and issue "starting" and "ending" messages.

We realize that this is a reversal (sorry tm Canada 🇨🇦)
of prior remarks about leaf Makefiles.  For those who object
to "starting" and "ending" messages they may set the
`${S}` makefile variable to `@#` as in:

```sh
make clobber S='@#'
```

And if you want all messages to go away and be "kept completely in the
dark" as the expression goes, then you could also do:

```sh
make clobber S=@# E=@#
```