Add is_available.sh
This script, which originally comes from the jparse repo
(https://github.com/xexyl/jparse), checks if a tool is available. This
is used instead of type -P in the Makefile.

For certain tools the script runs a trivial test to make sure it is
reliable and the Makefile now reports (for those certain tools) that it
cannot be found or is unreliable.