release 2.5.0 2024-09-26
Add `dyn_array_first(array, type)` helper macro.

Added `example.c` for an example of how to allocate a dynamic array,
store malloced string pointers, print strings, free storage.  This code is
presented as is: the `Makefile` does not build it as it also depends on
the `libdbg.a` and the `<dbg.h>` include file.  The `example` executable
is NOT needed as `example.c` just serves as an example, AND `dyn_test`
already does a proper job of testing.

Improved man page in various ways, including an expanded `RETURN VALUE`
and an expanded `EXAMPLE` section.  Added lots of warning about not
dereferencing the element pointer returned by `dyn_array_beyond()`
and caution about dereferencing the element pointer returned by
`dyn_array_first()`.

Simplified the `SYNOPSIS`, removing what the convenience macros
evaluate to as that expression is complex, requires that expression
to remain in sync with what is contained in `dyn_test.h`,
and is not needed for the synopsis.

Adopt section 3 style `SYNOPSIS` for functions: removed `extern`,
start with a return type preceding line, and a blank line between
functions.

Added `dyn_array_first(3)` man page copy of the main page.

Added a fun `BUGS` section to the man page.

Updated `man/man3/dyn_array.3` to include the `example.c` source.

Added a **PLEASE NOTE** about why `dyn_test.c` had to "import" code from
dbg to make it standalone to the source code, man page and `README.md`.

Updated `.gitignore` to ignore any building of `example`.

Added `example.c` to the `LESS_PICKY_CSRC` section of the `Makefile`.

Improved example links in `README.md`.

Added support to treat a dynamic array as if it was a LIFO (Last in,
First out) or "stack".  The `dyn_array_push()` macro pushes into
the stack, the `dyn_array_top()` reads the top of stack, and
the `dyn_array_pop()` pops he stack.

Updated the man pages with info about treating dynamic arrays as a "stack".

Added `dyn_array_pop(3)`, `dyn_array_push(3)`, and `dyn_array_top(3)`
as man page copies.

Changed `DYN_ARRAY_VERSION` from "2.4.0 2024-09-08"
to "2.5.0 2024-09-26".

This change is backward compatible with the dynamic array API.  No changes
to the previous API was made in this change.  Only new functionality
was added.  As such, code that already uses the dynamic array facility
does NOT need to change.

Ran `make clobber picky seqcexit all test` to test the above.